+++ src/clib/xtg/x_ijk2ib.c
@@ -4,7 +4,7 @@
 #include "libxtg_.h"
 
 long
-x_ijk2ib(int i, int j, int k, int nx, int ny, int nz, int ia_start)
+x_ijk2ib(long i, long j, long k, long nx, long ny, long nz, int ia_start)
 {
 
     if (i > nx || j > ny || k > nz) {
@@ -13,8 +13,8 @@ x_ijk2ib(int i, int j, int k, int nx, int ny, int nz, int ia_start)
         return -2;
     }
 
-    long ib = ((long)k - 1) * (long)nx * (long)ny;
-    ib = ib + ((long)j - 1) * (long)nx;
+    long ib = (k - 1) * nx * ny;
+    ib = ib + (j - 1) * nx;
     ib = ib + i;
 
     if (ia_start == 0)
@@ -25,7 +25,7 @@ x_ijk2ib(int i, int j, int k, int nx, int ny, int nz, int ia_start)
 
 /* c order counting, where K is looping fastest, them J, then I */
 long
-x_ijk2ic(int i, int j, int k, int nx, int ny, int nz, int ia_start)
+x_ijk2ic(long i, long j, long k, long nx, long ny, long nz, int ia_start)
 {
 
     if (i > nx || j > ny || k > nz) {
@@ -34,10 +34,9 @@ x_ijk2ic(int i, int j, int k, int nx, int ny, int nz, int ia_start)
         return -2;
     }
 
-
-    long ic = ((long)i - 1) * (long)nz * (long)ny;
-    ic = ic + ((long)j - 1) * nz;
-    ic = ic + (long)k;
+    long ic = (i - 1) * nz * ny;
+    ic = ic + (j - 1) * nz;
+    ic = ic + k;
 
     if (ia_start == 0)
         ic--;
